requirements.txt
Flask==3.0.2
tensorflow==2.15.0
keras==2.15.0
pillow==10.1.0
numpy==1.26.2
requests==2.31.0
Keras-Preprocessing==1.1.2